admin/requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
flask-cors=4.0.0
# pip install -U flask-cors
python-dotenv==1.0.0
werkzeug~=2.3.7

orator~=0.9.9
pymysql==1.1.0